Refactor Project to Support Net Standard
- Cleaned up namespaces
- Added Appveyor.yml for building + artifacting
- Added development dependency of GitVersion which programatically
creates the version based on git commits + tags. It also will put version
information in the assemblies generated
- Added Cake Build for Appveyor
- Removed Example project, the readme + unit tests should be sufficient
- Converted readme to pure Github Flavored Markdown
- Added in some sleeps in unit tests, the build server was occasionally not
acquiring the semaphore lock.
