+++ .editorconfig
@@ -1,0 +1,18 @@
+# EditorConfig helps developers define and maintain consistent
+# coding styles between different editors and IDEs
+# editorconfig.org
+
+root = true
+
+
+[*]
+indent_style = space
+indent_size = 4
+trim_trailing_whitespace = true
+insert_final_newline = false
+
+[*.{config,xml,js,json,html,css,sql,csproj,yml}]
+indent_size = 2
+
+[*.md]
+trim_trailing_whitespace = false
