+++ .markdownlint.json
@@ -1,0 +1,8 @@
+{
+  "default": true,
+  "MD003": { "style": "setext_with_atx_closed" },
+  "MD007": { "indent": 2 },
+  "MD013": true,
+  "no-hard-tabs": false,
+  "MD032": false
+}
