--- Example/ObjectBinder.cs
@@ -1,38 +1,0 @@
-﻿/*
-Copyright 2012, 2013, 2017 Adam Carter (http://adam-carter.com)
-
-This file is part of FileCache (http://github.com/acarteas/FileCache).
-
-FileCache is distributed under the Apache License 2.0.
-Consult "LICENSE.txt" included in this package for the Apache License 2.0.
-*/
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Reflection;
-
-namespace Example
-{
-    /// <summary>
-    /// You should be able to copy & paste this code into your local project to enable caching custom objects.
-    /// </summary>
-    public sealed class ObjectBinder : System.Runtime.Serialization.SerializationBinder
-    {
-        public override Type BindToType(string assemblyName, string typeName)
-        {
-            Type typeToDeserialize = null;
-
-            String currentAssembly = Assembly.GetExecutingAssembly().FullName;
-
-            // In this case we are always using the current assembly
-            assemblyName = currentAssembly;
-
-            // Get the type using the typeName and assemblyName
-            typeToDeserialize = Type.GetType(String.Format("{0}, {1}",
-                typeName, assemblyName));
-
-            return typeToDeserialize;
-        }
-    }
-}
