--- FileCache.UnitTests/CustomObjA.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace FC.UnitTests
-{
-    [Serializable]
-    public class CustomObjA
-    {
-        public string Name { get; set; }
-    }
-}
