--- FileCache.UnitTests/CustomObjB.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace FC.UnitTests
-{
-    [Serializable]
-    public class CustomObjB
-    {
-        public CustomObjA obj { get; set; }
-        public int Num { get; set; }
-    }
-}
