--- FileCache.sln
@@ -1,78 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.16
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileCache", "FileCache\FileCache.csproj", "{E13FD5FE-7D45-4A90-8FBB-C498BD144940}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileCache.Signed", "FileCache.Signed\FileCache.Signed.csproj", "{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileCache.UnitTests", "FileCache.UnitTests\FileCache.UnitTests.csproj", "{F7C96C00-C7CD-4503-B6BE-B79700D24F68}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3E97ABB4-233D-4500-85E2-53583EB0CF3B}"
-	ProjectSection(SolutionItems) = preProject
-		FileCache.vsmdi = FileCache.vsmdi
-		Local.testsettings = Local.testsettings
-		TraceAndTestImpact.testsettings = TraceAndTestImpact.testsettings
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Example", "Example\Example.csproj", "{3E796953-08A9-4B32-991D-9D7E187BA56D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|x86.ActiveCfg = Release|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|x86.ActiveCfg = Release|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|x86.ActiveCfg = Release|Any CPU
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Debug|x86.ActiveCfg = Debug|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Debug|x86.Build.0 = Debug|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Release|Any CPU.ActiveCfg = Release|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Release|Mixed Platforms.Build.0 = Release|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Release|x86.ActiveCfg = Release|x86
-		{3E796953-08A9-4B32-991D-9D7E187BA56D}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = FileCache.vsmdi
-	EndGlobalSection
-EndGlobal
