+++ src/FileCache.sln
@@ -1,0 +1,62 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.16
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileCache", "FileCache\FileCache.csproj", "{E13FD5FE-7D45-4A90-8FBB-C498BD144940}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileCache.Signed", "FileCache.Signed\FileCache.Signed.csproj", "{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileCache.UnitTests", "FileCache.UnitTests\FileCache.UnitTests.csproj", "{F7C96C00-C7CD-4503-B6BE-B79700D24F68}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{E13FD5FE-7D45-4A90-8FBB-C498BD144940}.Release|x86.ActiveCfg = Release|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{5BAF5DCF-5DE1-4806-B7E3-7EF0894A09E6}.Release|x86.ActiveCfg = Release|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{F7C96C00-C7CD-4503-B6BE-B79700D24F68}.Release|x86.ActiveCfg = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C5C8A797-9514-431C-857E-BCC3A16D3E94}
+	EndGlobalSection
+	GlobalSection(TestCaseManagementSettings) = postSolution
+		CategoryFile = FileCache.vsmdi
+	EndGlobalSection
+EndGlobal
